+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Added
+- Static method to obtain required enketo-transformer version (Form.getRequiredTransformerVersion)
+
 [4.16.2] - 2016-12-14
 ---------------------
 ##### Fixed